Refactor: Add new slash commands for quiz generation and slide summarization
This commit adds two new slash commands, /quizmode and /slidesummarizer, to the bot. The /quizmode command allows users to generate a quiz from a file, specifying the number of questions and the type of quiz (MCQ or Essay). The /slidesummarizer command allows users to summarize a slide deck by uploading the file. These new commands enhance the functionality of the bot and provide more options for users.